Interview Test Submission
Created the class BinFile to represent a binary file in code to do
comparison on. Overrode the Equals method to determine whether or not
two BinFile objects are equal according to the spec.

Created FileReader class to read binary files and create a BinFile
object from the read binary file.